Adding support for variable (aka field) expressions in smart predicates
Prior to this commit, we only supported generating smart predicates from expression that only used constant values. This commit implements support for FieldExpression by compiling a lambda that returns the value of the expression and invoking it at runtime. If the value returned is the same as the value that was passed into the function call, we assume that the predicate should return true.