Implementing support for Any() and Matches() smart predicates
When making setups for mocks, we now can specify that a parameter can have any value by using `MocksanePredicates.Any<TParameter>()` in our setup expression.

Similarly, we can define a smart expression that matches a simple `Func<TParameter, bool>` by using `Matches<TParamater>()` to prevent the need to write a complicated `predicateFunction` when dealing with a simple matcher.

These are the building blocks of ensuring that you can mock methods with a fair amount of flexibility.